Move part of logic into a separate package
