+++ packages/graphql-codegen-cli/package.json
@@ -55,6 +55,7 @@
     "common-tags": "1.8.0",
     "detect-indent": "5.0.0",
     "glob": "7.1.3",
+    "graphql-codegen-core": "0.18.0",
     "graphql-codegen-plugin-helpers": "0.18.0",
     "graphql-config": "2.2.1",
     "graphql-import": "0.7.1",
@@ -89,7 +90,8 @@
   },
   "main": "dist/commonjs/index.js",
   "module": "dist/esnext/index.js",
+  "typings": "dist/index.d.ts",
   "typescript": {
-    "definition": "./dist/index.d.ts"
+    "definition": "dist/index.d.ts"
   }
 }
