+++ packages/graphql-codegen-core/.gitignore
@@ -1,0 +1,6 @@
+node_modules
+npm-debug.log
+dist
+temp
+yarn-error.log
+tests/coverage/
