+++ packages/graphql-codegen-core/README.md
@@ -1,0 +1,9 @@
+# GraphQL Code Generator
+
+<p align="center">
+    <img src="https://github.com/dotansimha/graphql-code-generator/blob/master/logo.png?raw=true" />
+</p>
+
+Live demo and full documentation: [graphql-code-generator.com](https://graphql-code-generator.com)
+
+Project repository: [graphql-code-generator](https://github.com/dotansimha/graphql-code-generator)
