+++ packages/graphql-codegen-core/cjs/index.js
@@ -1,0 +1,1 @@
+module.exports = require('esm')(module, { force: true })('../dist/index.js');
