+++ packages/graphql-codegen-core/jest.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  globals: {
+    'ts-jest': {
+      tsConfig: {
+        esModuleInterop: true
+      }
+    }
+  }
+};
