+++ packages/graphql-codegen-core/package.json
@@ -1,0 +1,56 @@
+{
+  "name": "graphql-codegen-core",
+  "version": "0.18.0",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/dotansimha/graphql-code-generator.git"
+  },
+  "scripts": {
+    "clean": "rimraf ./dist",
+    "lint": "tslint src/**/*.ts",
+    "prebuild": "yarn clean",
+    "build": "tsc",
+    "test": "jest --no-cache --verbose --runInBand"
+  },
+  "keywords": [
+    "gql",
+    "generator",
+    "code",
+    "types",
+    "interfaces",
+    "graphql",
+    "codegen",
+    "apollo",
+    "node",
+    "typescript",
+    "ts",
+    "flow",
+    "types",
+    "d.ts",
+    "typings"
+  ],
+  "author": "Dotan Simha <dotansimha@gmail.com>",
+  "bugs": {
+    "url": "https://github.com/dotansimha/graphql-codegen/issues"
+  },
+  "homepage": "https://github.com/dotansimha/graphql-codegen#readme",
+  "peerDependencies": {
+    "graphql": "^0.8.0 || ^0.9.0 || ^0.10.0 || ^0.11.0 || ^0.12.0 || ^0.13.0 || ^14.0.0"
+  },
+  "dependencies": {
+    "esm": "3.2.11",
+    "graphql-codegen-plugin-helpers": "0.18.0",
+    "graphql-toolkit": "0.2.5",
+    "tslib": "1.9.3"
+  },
+  "devDependencies": {
+    "ts-jest": "24.0.0"
+  },
+  "main": "cjs/index.js",
+  "module": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "typescript": {
+    "definition": "dist/index.d.ts"
+  }
+}
