+++ packages/graphql-codegen-core/src/index.ts
@@ -1,0 +1,4 @@
+export * from './errors';
+export { codegen } from './codegen';
+export { executePlugin, ExecutePluginOptions } from './execute-plugin';
+export { mergeSchemas } from './merge-schemas';
